added remove comment

+++ client/src/actions/post.js
@@ -172,22 +172,26 @@ export const addComment = (postId, formData) => async dispatch => {
 
 // Delete comment
 export const deleteComment = (postId, commentId) => async dispatch => {
-  try {
-    await axios.delete(`/api/posts/comment/${postId}/${commentId}`)
+  if (
+    window.confirm('Are you sure you want to delete the comment permanently?')
+  ) {
+    try {
+      await axios.delete(`/api/posts/comment/${postId}/${commentId}`)
 
-    dispatch({
-      type: REMOVE_COMMENT,
-      payload: commentId
-    })
+      dispatch({
+        type: REMOVE_COMMENT,
+        payload: commentId
+      })
 
-    dispatch(setAlert('Comment Removed', 'success'))
-  } catch (error) {
-    dispatch({
-      type: POST_ERROR,
-      payload: {
-        msg: error.response.statusText,
-        status: error.response.status
-      }
-    })
+      dispatch(setAlert('Comment Removed', 'success'))
+    } catch (error) {
+      dispatch({
+        type: POST_ERROR,
+        payload: {
+          msg: error.response.statusText,
+          status: error.response.status
+        }
+      })
+    }
   }
 }
